Add variables to override load balancer's container name
Add `alb_container_name` and `nlb_container_name` variables, which are
prioritized when building the ECS load balancers, instead of using the
container name of the generated container definition.

An example use case for this implementation is adding a NGINX reverse
proxy container acting as a sidecar. For that configuration, the main
container definition is actually related to the downstream application
(e.g. Django application exposing a Gunicorn server in port 5000).
NGINX reverse proxy is configured within `init_containers`, exposing
port 80. Now, we need load balancers to point to the NGINX container in
port 80.

Up until now, it wasn't possible to have this configuration, as load
balancers would point to the Gunicorn server instead.