misc: Allow setting init_containers with null condition (#289)
This change allows configuring `init_containers` that are not added to
the `depends_on` list for the main container, by setting `null` as the
container `condition`.

This is useful for containers that actually depend on the main
container, and not the other way around. For example, a `nginx` sidecar
that needs to wait for the main container to be healthy before starting.

By allowing `null` as the container condition, this is a non-breaking
change.